+++ apps/web/app/api/comments/[commentId]/route.ts
@@ -1,2 +1,2 @@
 export const dynamic = 'force-dynamic';
-export { deleteComment as DELETE, updateComment as PUT } from '@/features/comment/server/comment.controller.';
+export { deleteComment as DELETE, updateComment as PUT } from '@/features/comment/server/comment.controller';
