+++ apps/web/app/api/comments/route.ts
@@ -1,2 +1,2 @@
 export const dynamic = 'force-dynamic';
-export { getPostComments as GET, createComment as POST } from '@/features/comment/server/comment.controller.';
+export { getPostComments as GET, createComment as POST } from '@/features/comment/server/comment.controller';
